Step 1 of attempting to resolve circular dependency with settings
