+++ cityhall/api/db/auth.py
@@ -12,11 +12,10 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-# >>>>>>>>>>>>>>>>>>>>
-# from django.conf import settings
+import cityhall.settings as settings
 from api.cache import CacheDict
 from api.db.env import Env
-from api.db.constants import Rights
+from api.db import Rights
 
 
 class Auth(object):
@@ -24,7 +23,7 @@ class Auth(object):
         self.db = db
         self.name = name
         self.roots_cache = CacheDict(
-            capacity=30 #settings.CACHE_OPTIONS['ENV_CAPACITY']
+            capacity=settings.CACHE_OPTIONS['ENV_CAPACITY']
         )
         self.user_root = user_root
         self.users_env = None
