+++ cityhall/api/db/django/db_factory.py
@@ -1,4 +1,4 @@
-from api.db.constants import DbFactory
+from api.db import DbFactory
 from api.db.django.db import Db
 from api.models import User, Value
