+++ cityhall/api/db/memory/db.py
@@ -12,7 +12,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from api.db.constants import Db
+from api.db import Db
 from datetime import datetime
