+++ cityhall/test/test_auth.py
@@ -13,7 +13,8 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from django.test import TestCase
-from api.db import Connection, Rights
+from api.db import Rights
+from api.db.connection import Connection
 from api.db.memory.db_factory import CityHallDbFactory
